Split PlantGrowthChamber specs into a package for readability.
Move observation/action specs from monolithic specs.py into focused
modules: base, simple_observations, day_area_trace, wall_stats, actions,
and registry. Public imports unchanged via specs/__init__.py.
